Remove extensions, move db global to models
